probleme de message negatif regle

+++ image.cpp
@@ -446,34 +446,34 @@ std::vector<ConfusionSet> Image::assignInitialPriority(const std::vector<Node>& 
 }*/
 
 
-int Image::ssdOverlap(Point n1, Point n2, Point p1, Point p2, int patchSize) const {
-    int ssd = 0;
+double Image::ssdOverlap(Point n1, Point n2, Point p1, Point p2, int patchSize) const {
+    double ssd = 0;
 
     if (n1.first +patchSize/2==n2.first) { //node 1 on the left
         for (int x = -patchSize/2; x <= 0; x++) {
             for (int y = -patchSize/2; y <= patchSize/2; y++) {
-                int diff = (*this)(p1.first +patchSize/2 + x, p1.second + y) - (*this)(p2.first + x, p2.second + y);
+                double diff = (*this)(p1.first +patchSize/2 + x, p1.second + y) - (*this)(p2.first + x, p2.second + y);
                 ssd += diff * diff;
             }
         }
     } else if (n2.first +patchSize/2 == n1.first) {//node 2 on the left
         for (int x = -patchSize/2; x <= 0; x++) {
             for (int y = -patchSize/2; y <= patchSize/2; y++) {
-                int diff = (*this)(p1.first + x, p1.second + y) - (*this)(p2.first + patchSize/2 + x, p2.second + y);
+                double diff = (*this)(p1.first + x, p1.second + y) - (*this)(p2.first + patchSize/2 + x, p2.second + y);
                 ssd += diff * diff;
             }
         }
     } else if (n1.second+patchSize/2 == n2.second) {//n1 on the bottom
         for (int x = -patchSize/2; x <= patchSize/2; x++) {
             for (int y = 0; y < patchSize/2 + 1; y++) {
-                int diff = (*this)(p1.first + x, p1.second + y) - (*this)(p2.first + x, p2.second - patchSize/2 - 1 + y);
+                double diff = (*this)(p1.first + x, p1.second + y) - (*this)(p2.first + x, p2.second - patchSize/2 - 1 + y);
                 ssd += diff * diff;
             }
         }
     } else if (n2.second+patchSize/2 == n1.second){//n2 on the bottom
         for (int x = -patchSize/2; x <= patchSize/2; x++) {
             for (int y = -patchSize/2; y <= 0; y++) {
-                int diff = (*this)(p1.first + x, p1.second + y) - (*this)(p2.first + x, p2.second + patchSize/2 + y);
+                double diff = (*this)(p1.first + x, p1.second + y) - (*this)(p2.first + x, p2.second + patchSize/2 + y);
                 ssd += diff * diff;
             }
         }
